requirements.txt
blinker==1.4
cachelib==0.4.1
click==8.0.3
Flask==2.0.2
Flask-Mail==0.9.1
Flask-Session==0.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
python-dotenv==0.19.2
python-http-client==3.3.5
sendgrid==6.9.4
starkbank-ecdsa==2.0.3
Werkzeug==2.0.2
